+++ ccip-sdk/src/evm/abi/OffRamp_2_0.ts
@@ -222,9 +222,13 @@ export default [
     name: 'getAllSourceChainConfigs',
     inputs: [],
     outputs: [
-      { name: '', type: 'uint64[]', internalType: 'uint64[]' },
       {
-        name: '',
+        name: 'sourceChainSelectors',
+        type: 'uint64[]',
+        internalType: 'uint64[]',
+      },
+      {
+        name: 'sourceChainConfigs',
         type: 'tuple[]',
         internalType: 'struct OffRamp.SourceChainConfig[]',
         components: [
@@ -420,25 +424,6 @@ export default [
   },
   {
     type: 'event',
-    name: 'MaxGasBufferToUpdateStateUpdated',
-    inputs: [
-      {
-        name: 'oldMaxGasBufferToUpdateState',
-        type: 'uint32',
-        indexed: false,
-        internalType: 'uint32',
-      },
-      {
-        name: 'newMaxGasBufferToUpdateState',
-        type: 'uint32',
-        indexed: false,
-        internalType: 'uint32',
-      },
-    ],
-    anonymous: false,
-  },
-  {
-    type: 'event',
     name: 'OwnershipTransferRequested',
     inputs: [
       {
@@ -610,6 +595,19 @@ export default [
     type: 'error',
     name: 'InvalidEncodingVersion',
     inputs: [{ name: 'version', type: 'uint8', internalType: 'uint8' }],
+  },
+  {
+    type: 'error',
+    name: 'InvalidFinalityForReceiver',
+    inputs: [
+      { name: 'receiver', type: 'address', internalType: 'address' },
+      { name: 'msgBlockDepth', type: 'uint16', internalType: 'uint16' },
+      {
+        name: 'requiredBlockDepth',
+        type: 'uint16',
+        internalType: 'uint16',
+      },
+    ],
   },
   {
     type: 'error',